==> preprocessing/2b_Mouse_LabelGABAergic.Rmd <==
---
title: "2b. Mouse, Label GABAergic"
output: 
---

```{r setup, include=FALSE}
# This sets the project root based on the repo structure.
# If you move this file, you may need to set the root manually to find config.R
knitr::opts_knit$set(root.dir = dirname(dirname(rstudioapi::getSourceEditorContext()$path)))
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

```{r libraries}
library(Seurat)
library(SeuratDisk)
library(reticulate)
library(scrubletR)
library(ggplot2)
library(patchwork)
library(dplyr)
library(data.table)
library(clustree)
library(reshape2)
library(tidyr)
library(comparatome)
source("config.R")
```

```{r load_data}
# Load preprocessed GABAergic cells
obj.mouse.gabaergic <- readRDS(paste0(dir.list$central$snrna$seurat$raw, "mouse_v1_gabaergic_raw.rds"))
```

```{r initial_clustering}
# Cluster at multiple resolutions
obj.mouse.gabaergic <- ClusterWithSCT(obj.mouse.gabaergic, c(1, 1.5, 2))
```

```{r explore_clustering}
# Examine clustering and metadata
DimPlot(obj.mouse.gabaergic, reduction = "umap", group.by = "sample", label = FALSE, raster = FALSE, shuffle = TRUE) + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.gabaergic, reduction = "umap", group.by = "subclass", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.gabaergic, reduction = "umap", group.by = "type", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.gabaergic, reduction = "umap", group.by = "SCT_snn_res.1", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.gabaergic, reduction = "umap", group.by = "SCT_snn_res.1.5", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.gabaergic, reduction = "umap", group.by = "SCT_snn_res.2", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
```

```{r canonical_markers}
# Define and visualize canonical marker genes
canon.markers <- list(
                    class = c("Gad1", "Gad2"),
                    Pvalb = c("Pvalb", "Myo5b", "Cemip"), 
                    Sst = c("Sst", "Chodl"), 
                    Vip = c("Vip"), 
                    Lamp5 = c("Lamp5"), 
                    Other = c("Frem1", "Stac", "Sncg")
                      )

DefaultAssay(obj.mouse.gabaergic) <- "RNA"
Idents(obj.mouse.gabaergic) <- "type"
PlotFeatures(obj.mouse.gabaergic, canon.markers)
DotPlot(obj.mouse.gabaergic, features = canon.markers, cols = c("lightgrey", "red"), scale = FALSE) + 
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.25)) + 
                         theme(panel.background = element_rect(fill = "white", color = NA), plot.background = element_rect(fill = "white", color = NA)) + NoLegend()
```

```{r clustree}
# Visualize clustering tree across resolutions
clustree(obj.mouse.gabaergic, prefix = "SCT_snn_res.")
```

```{r assign_subclasses}
# Map existing types and clusters to subclasses
subclass.idx.mouse <- list()

subclass.idx.mouse$type$Pvalb <- c("Pvalb_A", "Pvalb_B", "Pvalb_C", "Pvalb_D")
subclass.idx.mouse$SCT_snn_res.1$Pvalb <- c("2", "3", "7", "12")
subclass.idx.mouse$SCT_snn_res.1.5$Pvalb <- c("1", "3", "4", "7")
subclass.idx.mouse$SCT_snn_res.2$Pvalb <- c("1", "2", "6", "12", "13")

subclass.idx.mouse$type$Sst <- c("Sst_A", "Sst_B", "Sst_C", "Sst_D", "Sst_E")
subclass.idx.mouse$SCT_snn_res.1$Sst <- c("5", "6", "8", "9")
subclass.idx.mouse$SCT_snn_res.1.5$Sst <- c("5", "6", "8", "13", "14")
subclass.idx.mouse$SCT_snn_res.2$Sst <- c("3", "5", "10", "14", "15", "18")

subclass.idx.mouse$type$Vip <- c("Vip_A", "Vip_B", "Vip_C")
subclass.idx.mouse$SCT_snn_res.1$Vip <- c("1", "11")
subclass.idx.mouse$SCT_snn_res.1.5$Vip <- c("10", "11", "12")
subclass.idx.mouse$SCT_snn_res.2$Vip <- c("8", "9", "11")

subclass.idx.mouse$type$Lamp5 <- c("Lamp5")
subclass.idx.mouse$SCT_snn_res.2$Lamp5 <- c("4", "7")

subclass.idx.mouse$type$Frem1 <- c("Frem1")
subclass.idx.mouse$type$Stac <- c("Stac")

obj.mouse.gabaergic <- SubclassByIdent(obj.mouse.gabaergic, subclass.idx.mouse)
```

```{r plot_proportions}
# Visualize relative subclass proportions across samples
Idents(obj.mouse.gabaergic) <- "subclass"
levels(obj.mouse.gabaergic) <- c("Pvalb", "Sst", "Vip", "Lamp5", "Frem1", "Stac")

IdentBySample(obj.mouse.gabaergic)
```

```{r type_markers_final}
# Compute markers for final type assignments
subclass.labels <- c("Pvalb", "Sst", "Vip")
ident.labels <- c("type")

subclass.markers.mouse.gabaergic <- SubclassMarkerDict(obj.mouse.gabaergic, "subclass",
                                                       paste0(dir.list$pre$data, "mouse/gabaergic/markerdict_subclass.rds"))
markers.mouse.gabaergic <- IdentMarkerDict(obj.mouse.gabaergic, subclass.labels, ident.labels,
                                        paste0(dir.list$pre$data, "mouse/gabaergic/markerdict_types.rds"))
SaveDotPlots(obj.mouse.gabaergic, markers.mouse.gabaergic,
             subclass.labels, ident.labels,
             paste0(dir.list$pre$plots, "mouse/gabaergic/"), "ENSMUSG")
SaveFeaturePlots(obj.mouse.gabaergic, markers.mouse.gabaergic,
                 subclass.labels, ident.labels,
                 paste0(dir.list$pre$plots, "mouse/gabaergic/"))

PlotSubclassGeneCounts(subclass.markers.mouse.gabaergic, "subclass", 
                       c("Pvalb", "Sst", "Vip", "Lamp5", "Frem1", "Stac")) + theme(aspect.ratio = 1)
for (sbcl in subclass.labels) {
  for (id in ident.labels) {
    p <- PlotIdentGeneCounts(markers.mouse.gabaergic, sbcl, id)
    p <- p + theme(aspect.ratio = 1)
    print(p)
  }
}

SaveSubclassConfusionMatrices(obj.mouse.gabaergic, "subclass",
                              c("Pvalb", "Sst", "Vip", "Lamp5", "Frem1", "Stac"),
                              paste0(dir.list$pre$plots, "mouse/gabaergic/"))
SaveIdentConfusionMatrices(obj.mouse.gabaergic, subclass.labels, ident.labels,
                           paste0(dir.list$pre$plots, "mouse/gabaergic/"))
```

```{r save_processed}
# Save processed object with final labels
saveRDS(obj.mouse.gabaergic, paste0(dir.list$snrna$seurat$processed, "mouse_v1_gabaergic_processed.rds"))
```
