==> preprocessing/2c_Mouse_LabelNonneuronal.Rmd <==
---
title: "2c. Mouse, Label Nonneuronal"
output: 
---

```{r setup, include=FALSE}
# This sets the project root based on the repo structure.
# If you move this file, you may need to set the root manually to find config.R
knitr::opts_knit$set(root.dir = dirname(dirname(rstudioapi::getSourceEditorContext()$path)))
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

```{r libraries}
library(Seurat)
library(SeuratDisk)
library(reticulate)
library(scrubletR)
library(ggplot2)
library(patchwork)
library(dplyr)
library(data.table)
library(clustree)
library(reshape2)
library(tidyr)
library(comparatome)
source("config.R")
```

```{r load_data}
# Load preprocessed nonneuronal cells
obj.mouse.nonneuronal <- readRDS(paste0(dir.list$snrna$seurat$raw, "mouse_v1_nonneuronal_raw.rds"))
```

```{r initial_clustering}
# Cluster at multiple resolutions
obj.mouse.nonneuronal <- ClusterWithSCT(obj.mouse.nonneuronal, c(1, 1.5, 2))
```

```{r explore_clustering}
# Examine clustering and metadata
DimPlot(obj.mouse.nonneuronal, reduction = "umap", group.by = "sample", label = FALSE, raster = FALSE, shuffle = TRUE) + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.nonneuronal, reduction = "umap", group.by = "subclass", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.nonneuronal, reduction = "umap", group.by = "type", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.nonneuronal, reduction = "umap", group.by = "SCT_snn_res.1", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.nonneuronal, reduction = "umap", group.by = "SCT_snn_res.1.5", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
DimPlot(obj.mouse.nonneuronal, reduction = "umap", group.by = "SCT_snn_res.2", label = TRUE, raster = FALSE, shuffle = TRUE) + NoLegend() + xlim(-18, 18) + ylim(-18, 18) + coord_equal()
```

```{r canonical_markers}
# Define and visualize canonical marker genes
canon.markers <- list(
                    Class  = c("Snap25", "Sv2b", "Gad1"),
                    Astro = c("Aldh1l1", "Grin2c", "Phkg1"), 
                    Endo = c("Pecam1", "Mecom", "Flt1"), 
                    Micro = c("Cx3cr1", "Arhgap45", "Inpp5d"), 
                    Oligo = c("Enpp6", "Mog", "St18"), 
                    OPC = c("Lhfpl3", "Pdgfra", "Cacng4"),
                    VLMC = c("Slc6a13", "Slc47a1", "Bnc2")
                      )

DefaultAssay(obj.mouse.nonneuronal) <- "RNA"
Idents(obj.mouse.nonneuronal) <- "type"
PlotFeatures(obj.mouse.nonneuronal, canon.markers)
DotPlot(obj.mouse.nonneuronal, features = canon.markers, cols = c("lightgrey", "red"), scale = FALSE) + 
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.25)) + 
                         theme(panel.background = element_rect(fill = "white", color = NA), plot.background = element_rect(fill = "white", color = NA)) + NoLegend()
```

```{r clustree}
# Visualize clustering tree across resolutions
clustree(obj.mouse.nonneuronal, prefix = "SCT_snn_res.")
```

```{r assign_subclasses}
# Map existing types to subclasses
subclass.idx.mouse <- list()

subclass.idx.mouse$type$Astro <- c("Astro_A", "Astro_B")
subclass.idx.mouse$type$Endo <- c("Endo")
subclass.idx.mouse$type$Micro <- c("Micro")
subclass.idx.mouse$type$OD <- c("OD_A", "OD_B", "OD_C")
subclass.idx.mouse$type$OPC <- c("OPC_A", "OPC_B")
subclass.idx.mouse$type$VLMC <- c("VLMC_A", "VLMC_B")

obj.mouse.nonneuronal <- SubclassByIdent(obj.mouse.nonneuronal, subclass.idx.mouse)
```

```{r plot_proportions}
# Visualize relative subclass proportions across samples
Idents(obj.mouse.nonneuronal) <- "subclass"
levels(obj.mouse.nonneuronal) <- c("Astro", "Endo", "Micro", "OD", "OPC", "VLMC")

IdentBySample(obj.mouse.nonneuronal)
```

```{r type_markers_final}
# Compute markers for final type assignments
subclass.labels <- c("Astro", "Endo", "Micro", "OD", "OPC", "VLMC")
ident.labels <- c("type")

subclass.markers.mouse.nonneuronal <- SubclassMarkerDict(obj.mouse.nonneuronal, "subclass",
                                                       paste0(dir.list$pre$data, "mouse/nonneuronal/markerdict_subclass.rds"))
markers.mouse.nonneuronal <- IdentMarkerDict(obj.mouse.nonneuronal, subclass.labels, ident.labels,
                                        paste0(dir.list$pre$data, "mouse/nonneuronal/markerdict_types.rds"))
SaveDotPlots(obj.mouse.nonneuronal, markers.mouse.nonneuronal,
             subclass.labels, ident.labels,
             paste0(dir.list$pre$plots, "mouse/nonneuronal/"), "ENSMUSG")
SaveFeaturePlots(obj.mouse.nonneuronal, markers.mouse.nonneuronal,
                 subclass.labels, ident.labels,
                 paste0(dir.list$pre$plots, "mouse/nonneuronal/"))

PlotSubclassGeneCounts(subclass.markers.mouse.nonneuronal, "subclass", 
                       c("Astro", "Endo", "Micro", "OD", "OPC", "VLMC")) + theme(aspect.ratio = 1)
for (sbcl in subclass.labels) {
  for (id in ident.labels) {
    p <- PlotIdentGeneCounts(markers.mouse.nonneuronal, sbcl, id)
    p <- p + theme(aspect.ratio = 1)
    print(p)
  }
}

SaveSubclassConfusionMatrices(obj.mouse.nonneuronal, "subclass",
                              c("Astro", "Endo", "Micro", "OD", "OPC", "VLMC"),
                              paste0(dir.list$pre$plots, "mouse/nonneuronal/"))
SaveIdentConfusionMatrices(obj.mouse.nonneuronal, subclass.labels, ident.labels,
                           paste0(dir.list$pre$plots, "mouse/nonneuronal/"))
```

```{r save_processed}
# Save processed object with final labels
saveRDS(obj.mouse.nonneuronal, paste0(dir.list$snrna$seurat$processed, "mouse_v1_nonneuronal_processed.rds"))
```
